refactor: rename autostart
