+++ src-tauri/src/settings/autostart.rs
